requirement.txt
asyncio-nats-client==0.5.2
docopt==0.6.2
logzero==1.3.0
passlib==1.7.1
py==1.4.34
pyserial==3.4
pytest==3.2.1
pytest-asyncio==0.7.0
python-periphery==1.1.0
PyYAML==3.12
spidev==3.2
transitions==0.2.5
turing-chatroom-bus
